requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
cfgv==3.3.1
colorama==0.4.5
distlib==0.3.5
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
factory-boy==3.2.1
Faker==13.15.1
filelock==3.7.1
identify==2.5.3
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.9.1
nodeenv==1.7.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
virtualenv==20.16.3
